reduced duplication code by implementing Generic and Abstract service

+++ server/src/main/java/cz/muni/fi/pv243/mustech/dal/UserRepository.java
@@ -5,7 +5,7 @@ import org.apache.deltaspike.data.api.EntityRepository;
 import org.apache.deltaspike.data.api.Repository;
 
 @Repository
-public interface UserRepository extends EntityRepository<User, Integer> {
+public interface UserRepository extends EntityRepository<User, Long> {
 
     User findById(Long id);
     User findByEmail(String name);
